Make registerAbscence work with Hjärntorget
This changes teh registerAbscence api to use DateTime instead of Date,
mainly to be consistent with getSchedule that already used DateTime.

The main fixes that was missing is properly formatting the start and end
dates sent to the server and ensuring a `Referer` header is included
similar to what is sent in the browser.